Fix: fixed some refactor bugs
