+++ packages/ledger/vm/devnetvm/input.go
@@ -15,7 +15,7 @@ import (
 	"github.com/iotaledger/hive.go/typeutils"
 	"github.com/mr-tron/base58"
 
-	"github.com/iotaledger/goshimmer/packages/refactored/ledger/utxo"
+	"github.com/iotaledger/goshimmer/packages/ledger/utxo"
 )
 
 // region InputType ////////////////////////////////////////////////////////////////////////////////////////////////////
